rag_service/requirements.txt
# RAG Service - ChromaDB + Embeddings + Document Processing
chromadb>=0.4.22
ollama>=0.3.0
pypdf>=4.0.0
python-docx>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
supabase>=2.0.0
